requirements.txt
streamlit
openai-whisper
sounddevice
scipy
requests
lingua-language-detector
numpy
